requirements.txt
django==3.0.8
djangorestframework==3.11.0
djangorestframework_simplejwt==4.4.0